do host key verification.
there was already code to verify whether the signature after the
key exchange was valid, we just didn't check with the user (or a
file) whether the host key (fingerprint of the host key) really
belonged to that address.  now we do.  we ask the user on /dev/cons
when an address is unknown.  when an address is known & matches,
we accept it.  if we have a different fingerprint or key, we raise
an error.  if we have a different algorithm, we raise an error too.
in the future, we could see if the user requested the other host
key algorithm himself and treat the key as potentially valid.