merge key exchange handling.
sshlib's login function is now gone.  it was used to setup the
connection (key exchange & user auth), but the key exchange part
was already duplicated (it had to).  now we have a handshake function
that only reads the ssh server banner, then we start the ssh message
loop from appl/cmd/ssh.b.  that loop hands off some messages to
sshlib.

note that this is quite ugly, it will be a big state machine... it
is somewhat necessary as many types of messages may come in at any
time and need to be properly handled.