split message formatting/parsing to separate lib.  cleanup.
the new lib is sshfmt, used by sshlib, ssh, sftpfs, and sshpubkey.
sftpfs previously depended on sshlib, which has all the internals
sftp has nothing to do with.  this is a cleaner separation.

we now also send disconnect messages more often when we don't like
what's going on (instead of just closing the connection).  and we
no longer fail for most unrecognized messages (still some left),
as the rfc's require.