split sftp message parsing & packing into library, appl/lib/sftp.[bm].
this makes the code in appl/cmd/sftpfs.b cleaner, and will help
when writing an sftpsrv.b in the future.

this does not change any of the processing of styx & sftp messages
in sftpfs, that's next.

sftpfs does now have an option -D to print sftp message traces.
manual page mentions it too.

ssh man page is fixed to correct -s usage, and to mention the user@
part of the address parameter.